requirements.txt
asposestorage==1.0.2
beautifulsoup4==4.12.3
db==0.1.1
langchain==0.2.12
langchain_community==0.2.11
langchain_core==0.2.28
langchain_google_genai==1.0.8
langchain_text_splitters==0.2.2
pandas==2.2.2
preprocessing==0.1.13
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.30
streamlit==1.32.0
tiktoken==0.7.0
